Release version 2.3.35
Source: 61ba72d19f241ffc0c895f24b061a0c202fec94d

Reviewed-on: https://gerrit.bouffalolab.com/c/release/bouffalo_sdk/+/11224
Tested-by: zrrong <zrrong@bouffalolab.com>

+++ components/wireless/wifi4/manager/bl60x_wifi_driver/wifi_mgmr.h
@@ -187,7 +187,7 @@ typedef struct
 typedef struct wifi_mgmr_scan_item {
     uint32_t mode;
     uint32_t timestamp_lastseen;
-    uint16_t ssid_len;
+    int ssid_len;
     uint8_t channel;
     int8_t rssi;
     char ssid[32];
